PHmiClient/Converters/MultiValueConverters/OrConverter.cs
using System;
using System.Linq;
using System.Windows.Data;

namespace PHmiClient.Converters.MultiValueConverters
{
    public class OrConverter : IMultiValueConverter
    {
        public object Convert(object[] values, Type targetType, object parameter, System.Globalization.CultureInfo culture)
        {
            return values.Any(value => value as bool? == true);
        }

        public object[] ConvertBack(object value, Type[] targetTypes, object parameter, System.Globalization.CultureInfo culture)
        {
            throw new NotSupportedException();
        }
    }
}
